requirements.txt
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
flywheel-bids==0.8.4.dev2
flywheel-gear-toolkit==0.1.0
flywheel-sdk==12.4.0
future==0.18.2
idna==2.9
jsonschema==3.2.0
numpy==1.19.2
pandas==1.1.2
pathlib==1.0.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
urllib3==1.25.10
zipp==3.1.0
